requirements.txt
asgiref==3.6.0
Django==4.1.7
django-mssql==1.8
django-phone-field==1.8.1
djangorestframework==3.14.0
joblib==1.2.0
mssql-django==1.2
numpy==1.24.2
pandas==2.0.0
pybase64==1.2.3
pycryptodome==3.17
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
gunicorn